# data/swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "API for Fixthis",
      "version": "0.1.0",
      "title": "Fixthis",
      "termsOfService": ""
   },
   "basePath": "/api",
...
   "schemes": [
      "http"
   ],
   "paths": {
...
      "/user/uploadpic": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Upload profile pic",
            "description": "Api to upload user profile pic",
            "operationId": "user.uploadprofilepic",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "profilePic",
                  "type": "file",
                  "required": true,
                  "description": "Pic to upload"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/uploadcoverpic": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Upload cover pic",
            "description": "Api to upload user cover pic",
            "operationId": "user.uploadcoverpic",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "coverPic",
                  "type": "file",
                  "required": true,
                  "description": "Pic to upload"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/user/onlinestatus": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Set user online status",
            "description": "",
            "operationId": "user.setonlinestatus",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "status",
                  "type": "string",
                  "enum": [
                     "0",
                     "1"
                  ],
                  "required": true,
                  "description": "Set current user online status"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/changepassword": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "API to update user password",
            "description": "Api to change user password",
            "operationId": "api.changeuserpassword",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "current_password",
                  "type": "string",
                  "required": true,
                  "description": "Current User Password"
               },
               {
                  "in": "formData",
                  "name": "new_password",
                  "type": "string",
                  "required": true,
                  "description": "New User Password"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/user/sendotp": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Send OTP to user",
            "description": "Api to send OTP to user",
            "operationId": "user.sendotp",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "mobile",
                  "type": "string",
                  "required": true,
                  "description": "Phone number without country code"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/verifyotp": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "verify user OTP",
            "description": "Api to verify OTP send to user",
            "operationId": "user.verifyotp",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "otp",
                  "type": "string",
                  "required": true,
                  "description": "Received OTP"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
   },
...
   "securityDefinitions": {
      "bearer": {
         "in": "header",
         "name": "Authorization",
         "type": "apiKey"
      }
   }
}